marble: Unlock max prime cpu freq in default thermal profile
Allow prime cpu to go up to 2.92 GHz in default thermal profile too.
Don't depend on user to set high performance mode to be able to use the
full potential of the cpu, especially now that per-app thermal profiles
are allowed to be turned off completely.

Previous limit was absurd - max 2.59 GHz unless the skin temperature
was under 25 C, which is unlikely in most weather conditions.
With this change, max 2.92 GHz is allowed until the skin temperature
hits 37 C at which point it is limited to 2.59 GHz until it drops down
to 36 C. Rest of the triggers are kept intact.

Diff:
https://gist.github.com/adithya2306/a9b82f389d69f357edf8104e3cf4c7e6/revisions

Change-Id: Idd081c793bef96f00e2ac9f9ebc802b44d4d7b4c